Right Claws: the tide calls a claw and the beach has to sort
A ninth face on the sparkling crab's roulette. While it runs, a
right-clawed crab banks for double and a left-clawed one costs what it
was worth, so ten seconds of herding whatever is nearest turns into ten
seconds of sorting. The claw a crab leads with was the one thing that
made this game different from the one it is built on, and until now only
the campaign ever asked anyone to look at it.

Always the right claw, never drawn. The rule falls on the whole beach at
once, so a fixed one is as fair as a coin and is learned once instead of
read off the banner every time. Nothing is needed to read it either: a
crab already wears its claw in blue or yellow, on the side it leads with.

Down to nothing and no further. A score that has gone negative is a
player who has stopped playing, and this game is for children.

Both ways a crab could be banked went through their own `scores[owner] +=`
line, the walk into a castle and Crab Monopoly's sweep, so a rule added
to one would have missed the other: they share `credit_bank` now. The
bots read `bank_worth` before they go fetching, or they would spend the
event carrying their own points away.

This is a rules change, not an addition. The roulette picks with one
modulo over `TideEvent::ALL`, so a ninth face moves every draw from the
first spin: the determinism anchor is re-derived, and the replay header
goes to `replay-v2` so a recording made under eight faces is refused
rather than quietly played out as a different round. The seen-it mask
needed a ninth bit, and Full Roulette now asks for nine.

+++ src/app/achievements/mod.rs
@@ -34,7 +34,7 @@ pub struct Stats {
     pub giants: u32,
     /// One bit per tide event ever spun (see [`TideEvent::index`]), so the
     /// roulette trophy needs *variety* rather than volume.
-    pub events_seen: u8,
+    pub events_seen: u16,
     /// Most crabs banked in a single round.
     pub best_round: u32,
     /// Series taken, of either length.
